+++ app/Http/Controllers/SaleInvoiceController.php
@@ -2,13 +2,15 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\JobTitle;
+use App\Models\Employee;
 use App\Models\Product;
 use App\Models\SaleInvoice;
 use App\Models\Service;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Gate;
+
+use PDF;
 
 class SaleInvoiceController extends Controller
 {
@@ -23,9 +25,15 @@ class SaleInvoiceController extends Controller
      */
     public function index()
     {
-        if(Gate::allows('sale_invoices',Auth::user())){
-            $saleInvoices = SaleInvoice::all();
-            return view('sale_invoices.index',compact('saleInvoices'));
+        if(Gate::allows('sale_invoices',Auth::user())||Gate::allows('sale_invoices_edit',Auth::user())){
+            $todaySaleInvoices = SaleInvoice::query()->where('created_at','>',date('Y-m-d'))->get();
+            $appliedSaleInvoices = SaleInvoice::query()
+                ->where('state','=','1')->get();
+            $savedSaleInvoices = SaleInvoice::query()
+                ->where('state','=','0')->get();
+            $canceledSaleInvoices = SaleInvoice::query()
+                ->where('state','=','2')->get();
+            return view('sale_invoices.index',compact('todaySaleInvoices','appliedSaleInvoices','savedSaleInvoices','canceledSaleInvoices'));
         }else{
             abort(401);
         }
@@ -41,21 +49,19 @@ class SaleInvoiceController extends Controller
 
     public function create()
     {
-        if(!session()->exists('counter')){
-            session()->put('counter',1);
+        if(!Gate::allows('sale_invoices_create',Auth::user())){
+            abort(401);
         }
-        $invoicesCounter = session()->get('counter');
-        $latest = SaleInvoice::latest()->first();
+
+        $latest = SaleInvoice::query()->where('state','=','1')->orderByDesc('id')->first();
+        $counter = 1;
         if ($latest!=null){
-            if(date_format($latest['created_at'],'Y-m-d')!=date('Y-m-d')){
-                session()->put('counter',1);
+            if(date_format($latest['created_at'],'Y-m-d')==date('Y-m-d')) {
+                $counter =  $latest->counter+ 1;
             }
         }
-        $jobTitles = JobTitle::query()->where("use_system",'=','0')->get();
-        $employees=[];
-        foreach ($jobTitles as $jobTitle) {
-            array_push($employees,$jobTitle->employees);
-        }
+        $invoicesCounter =$counter;
+        $employees = Employee::query()->where('commission','>','0')->get();
 
         $products = Product::all();
 
@@ -72,10 +78,9 @@ class SaleInvoiceController extends Controller
      */
     public function store(Request $request)
     {
-        //
-        
-
-        session()->put('counter',session()->get('counter')+1);
+        if(!Gate::allows('sale_invoices_create',Auth::user())){
+            abort(401);
+        }
 
         $request->validate([
             'total'=>['numeric','min:1','required'],
@@ -83,33 +88,63 @@ class SaleInvoiceController extends Controller
         ]);
         $products=array();
         $services=array();
+        $subTotal=0;
         foreach ($request->all() as $requestValue){
             if(substr( $requestValue, 0, 3 ) === 'p_-' ){
                 $splited = explode('-',$requestValue);
                 array_push($products, ['product_id'=>$splited[1],'price'=>$splited[2],'quantity'=>$splited[3]]);
+                $subTotal+=$splited['2']*$splited[3];
             }else if(substr( $requestValue, 0, 3 ) === 's_-' ){
                 $splited = explode('-',$requestValue);
-                array_push($services, ['employee_id'=>$splited[1],'service_id'=>$splited[2],'price'=>$splited[3],'quantity'=>$splited[4]]);
+                $totalService = $splited[3]*$splited[4]-$splited[3]*$splited[4]*($request['discount']/100);
+                array_push($services, ['employee_id'=>$splited[1],'service_id'=>$splited[2],'price'=>$splited[3],'quantity'=>$splited[4],'after_discount'=>$totalService,'state'=>$request['state']]);
+                $subTotal+=$splited[3]*$splited[4];
             }
         }
-        $invoice = Auth::user()->saleInvoices()->create($request->all());
+        $total=$subTotal - $subTotal*($request['discount']/100);
+        if($request['total']!=$total){
+            return redirect()->back()->withErrors([__('Something went wrong!')]);
+        }
 
+        $invoice = Auth::user()->saleInvoices()->create($request->all()+['sub_total'=>$subTotal]);
+        $productsInvoice=[];
         foreach ($products as $product) {
             $sale = $invoice->saleProducts()->create($product);
-            $newQuantity = $sale->product['quantity'] - $sale['quantity'];
-            if ($newQuantity < 0) {
-                $invoice->delete();
-                redirect()->back()->with('error', 'Out of stock' . $sale->product['name']);
+            if($invoice['state']==1) {
+                $newQuantity = $sale->product['quantity'] - $sale['quantity'];
+                if ($newQuantity < 0) {
+                    $invoice->saleProducts()->delete();
+                    $invoice->saleServices()->delete();
+                    $invoice->delete();
+                    return redirect()->back()->withErrors([__('Out of stock') . ' \'' . $sale->product['name'] . '\' ' .
+                        __('Available quantity is') . ' ' . $sale->product['quantity']]);
+                }
+                $sale->product()->update(['quantity' => $newQuantity]);
             }
-            $sale->product()->update(['quantity' => $newQuantity]);
+            array_push($productsInvoice,['name'=>$sale->product['name'],'price'=>$sale['price'],'quantity'=>$sale['quantity']]);
+
         }
 
 
+        $servicesInvoice = [];
         foreach ($services as $service) {
             $sale = $invoice->saleServices()->create($service);
+            array_push($servicesInvoice, ['name'=>$sale->service['name'],'price'=>$sale['price'],'quantity'=>$sale['quantity'],'specialist'=>$sale->employee['name']]);
         }
-
+        if($request['state']==1) {
+            $counter = $request['counter'];
+            $this->print($counter, $servicesInvoice, $productsInvoice, $invoice);
+        }
         return redirect()->back();
+    }
+
+    private function print($counter,$servicesInvoice,$productsInvoice,$invoice){
+
+        $pdf = PDF::loadView('frontend.pdf',compact('servicesInvoice','productsInvoice','invoice','counter'));
+        $pdf->getMpdf()->charset_in='UTF-8';
+
+        $pdf->save('storage/saleInvoices/invoice'.$invoice['id'].'.pdf','F');
+        exec('lp -o fit-to-page '.'storage/saleInvoices/invoice'.$invoice['id'].'.pdf');
     }
 
     /**
@@ -120,7 +155,14 @@ class SaleInvoiceController extends Controller
      */
     public function show(SaleInvoice $saleInvoice)
     {
-        //
+
+        if(!Gate::allows('sale_invoices',Auth::user()) || !Gate::allows('sale_invoices_edit',Auth::user())){
+            abort(401);
+        }
+
+        $invoicePath = asset("storage/saleInvoices/invoice".$saleInvoice['id'].'.pdf');
+        return view('sale_invoices.view',compact('invoicePath'));
+
     }
 
     /**
@@ -131,11 +173,10 @@ class SaleInvoiceController extends Controller
      */
     public function edit(SaleInvoice $saleInvoice)
     {
-        $jobTitles = JobTitle::query()->where("use_system",'=','0')->get();
-        $employees=[];
-        foreach ($jobTitles as $jobTitle) {
-            array_push($employees,$jobTitle->employees);
+        if(!Gate::allows('sale_invoices_edit',Auth::user())){
+            abort(401);
         }
+        $employees = Employee::query()->where('commission','>','0')->get();
 
         $products = Product::all();
 
@@ -155,7 +196,9 @@ class SaleInvoiceController extends Controller
      */
     public function update(Request $request, SaleInvoice $saleInvoice)
     {
-
+        if(!Gate::allows('sale_invoices_edit',Auth::user())){
+            abort(401);
+        }
         $request->validate([
             'total'=>['numeric','min:1','required'],
             'customer' => ['max:255'],
@@ -163,14 +206,22 @@ class SaleInvoiceController extends Controller
 
         $products=array();
         $services=array();
+        $subTotal=0;
         foreach ($request->all() as $requestValue){
             if(substr( $requestValue, 0, 3 ) === 'p_-' ){
                 $splited = explode('-',$requestValue);
                 array_push($products, ['product_id'=>$splited[1],'price'=>$splited[2],'quantity'=>$splited[3]]);
+                $subTotal+= $splited[2];
             }else if(substr( $requestValue, 0, 3 ) === 's_-' ){
                 $splited = explode('-',$requestValue);
-                array_push($services, ['employee_id'=>$splited[1],'service_id'=>$splited[2],'price'=>$splited[3],'quantity'=>$splited[4]]);
+                $totalService = $splited[3]*$splited[4]-$splited[3]*$splited[4]*($request['discount']/100);
+                array_push($services, ['employee_id'=>$splited[1],'service_id'=>$splited[2],'price'=>$splited[3],'quantity'=>$splited[4],'after_discount'=>$totalService,'state'=>$request['state']]);
+                $subTotal+= $splited[3];
             }
+        }
+        $total=$subTotal - $subTotal*($request['discount']/100);
+        if($request['total']!=$total){
+            redirect()->back()->with('error',__('Something went wrong!'));
         }
         $productsInPurchases = $saleInvoice->saleProducts()->get();
 
@@ -182,22 +233,44 @@ class SaleInvoiceController extends Controller
         $saleInvoice->saleProducts()->delete();
         $saleInvoice->saleServices()->delete();
 
-
+        $productsInvoice=[];
         foreach ($products as $product) {
             $sale = $saleInvoice->saleProducts()->create($product);
-            $newQuantity = $sale->product['quantity'] - $sale['quantity'];
-            if ($newQuantity < 0) {
-                $sale->delete();
-                redirect()->back()->with('error', 'Out of stock' . $sale->product['name']);
+            if($saleInvoice['state']==1) {
+                $newQuantity = $sale->product['quantity'] - $sale['quantity'];
+                if ($newQuantity < 0) {
+                    $saleInvoice->saleProducts()->delete();
+                    $saleInvoice->saleServices()->delete();
+                    $saleInvoice->delete();
+                    return redirect()->back()->withErrors([__('Out of stock') . ' \'' . $sale->product['name'] . '\' ' .
+                        __('Available quantity is') . ' ' . $sale->product['quantity']]);
+                }
+                $sale->product()->update(['quantity' => $newQuantity]);
             }
-            $sale->product()->update(['quantity' => $newQuantity]);
+            array_push($productsInvoice,['name'=>$sale->product['name'],'price'=>$sale['price'],'quantity'=>$sale['quantity']]);
         }
 
-
+        $servicesInvoice=[];
         foreach ($services as $service) {
             $sale = $saleInvoice->saleServices()->create($service);
+            array_push($servicesInvoice, ['name'=>$sale->service['name'],'price'=>$sale['price'],'quantity'=>$sale['quantity'],'specialist'=>$sale->employee['name']]);
         }
-        $saleInvoice->update($request->all());
+        if($saleInvoice['counter']==0){
+            $latest = SaleInvoice::query()->where('state','=','1')->orderByDesc('id')->first();
+            $counter = 1;
+            if ($latest!=null){
+                if(date_format($latest['created_at'],'Y-m-d')==date('Y-m-d')) {
+                    $counter =  $latest->counter+ 1;
+                    $saleInvoice->update($request->all()+['sub_total'=>$subTotal,'counter'=>$counter]);
+                }
+            }
+        }else{
+            $saleInvoice->update($request->all()+['sub_total'=>$subTotal]);
+
+        }
+        if($request['state']==1){
+            $this->print($saleInvoice->counter,$servicesInvoice,$productsInvoice,$saleInvoice);
+        }
         return redirect()->back();
     }
 
@@ -209,16 +282,28 @@ class SaleInvoiceController extends Controller
      */
     public function destroy(SaleInvoice $saleInvoice)
     {
+        if(!Gate::allows('sale_invoices_delete',Auth::user())){
+            abort(401);
+        }
         $productsInPurchases = $saleInvoice->saleProducts()->get();
-
+        $servicesInPurchases = $saleInvoice->saleServices()->get();
+        foreach ($servicesInPurchases as $service){
+            $service->state = 2;
+            $service->save();
+        }
         for ($i=0;$i<count($saleInvoice->saleProducts);$i++){
             $product = $productsInPurchases[$i]->product;
             $product->quantity+=$productsInPurchases[$i]->quantity;
             $product->save();
         }
-        $saleInvoice->saleProducts()->delete();
-        $saleInvoice->saleServices()->delete();
-        $saleInvoice->delete();
+        if($saleInvoice->state==0){
+            $saleInvoice->saleServices()->delete();
+            $saleInvoice->saleProducts()->delete();
+            $saleInvoice->delete();
+        }else {
+            $saleInvoice->state = 2;
+            $saleInvoice->save();
+        }
         return redirect()->to("sale_invoices");
     }
 }
